refactor: use custom hook to get windowid
